@@ -3,10 +3,16 @@
   "version": "1.0.0",
   "description": "",
   "main": "server/server.js",
+  "type": "module",
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "body-parser": "^1.20.2",
     "bootstrap": "^5.3.3",
+    "cookie-parser": "^1.4.6",
+    "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "express-session": "^1.18.0",
+    "jsonwebtoken": "^9.0.2",
     "sharp": "^0.33.4"
   },
   "scripts": {
@@ -17,7 +23,7 @@
     "nodemon": "^3.1.1"
   },
   "engines": {
-    "node": "14.x"
+    "node": "20.x"
   },
   "author": "",
   "license": "ISC"
