@@ -50,7 +50,7 @@ document.addEventListener("DOMContentLoaded", async () => {
                     </div>
                   </div>
                   <div class="card_bottom_body">
-                    <a class="card_link" href="/offer-details?game=${pageName}&item=${tarjeta.id}">
+                    <a class="card_link" href="/store/offer-details?game=${pageName}&item=${tarjeta.id}">
                       <div class="fs-sm d-flex flex-row flex-lg-column gap-sm-2">
                         <div class="d-flex align-items-center h-100">
                           <span class="descuento_label p-1">${tarjeta.descuento}</span>
@@ -93,7 +93,7 @@ document.addEventListener("DOMContentLoaded", async () => {
                     </div>
                   </div>
                   <div class="card_bottom_body">
-                    <a class="card_link" href="/games-details?game=${pageName}&item=${tarjeta.id}">
+                    <a class="card_link" href="/store/games-details?game=${pageName}&item=${tarjeta.id}">
                       <div class="card_price">
                         <span class="card_text">Desde</span>
                         <span class="price">CLP$ ${tarjeta.precioOriginal}</span>
@@ -129,7 +129,7 @@ document.addEventListener("DOMContentLoaded", async () => {
                     </div>
                   </div>
                   <div class="card_bottom_body">
-                    <a class="card_link" href="/gift-details?game=${pageName}&item=${tarjeta.id}">
+                    <a class="card_link" href="/store/gift-details?game=${pageName}&item=${tarjeta.id}">
                       <div class="card_price">
                         <span class="card_text">Desde</span>
                         <span class="price">CLP$ ${tarjeta.precioOriginal}</span>
