@@ -103,7 +103,7 @@ function renderCards(cardsForLoad, cardsContainer) {
             </div>
           </div>
           <div class="card_bottom_body">
-            <a class="card_link" href="/games-details?game=${pageName}&item=${tarjeta.id}">
+            <a class="card_link" href="/store/games-details?game=${pageName}&item=${tarjeta.id}">
               <div class="card_price">
                 <span class="card_text">Desde</span>
                 <span class="price">CLP$ ${tarjeta.precioOriginal}</span>
